Fixed profit calculations and item list
Profit calculations now only show profits for items when both source and destination prices are above 0 and the profit is positive (excludes items whose prices have not been set and unprofitable trades)

The list of trade goods is now pulled from a JSON file which can be edited to add more items (for mods, &c). If no file is detected at program start, the list is restored from an internal backup list of default trade items.